fix: handle disconnected revision logs
This change fixes an issue where revsets like `visible_heads()` return a
list of disconnected logs that are separated by an empty line and the
jjui log parser stopped parsing after it reached an empty line. As a
result of this, only the first revision of such revsets was displayed
previously.