fix: determine change and commit id by likelihood
Instead of finding ids by relying on the rest block, find a segment such
that if it looks like a change id then it is probably a change id,
similary if it is hex like then it is probably a commit id. Seems to
work in general case for now.

#40