feat(revisions): load graph rows in batches
This change updates the log parser to return parsed rows in batches and
wait until a new batch is requested. Initial batch size is set to 50. A
new batch is requested whenever the user moves the cursor beyond the
last loaded row. 

This change improves the responsiveness of `jjui` when working with huge
repositories while reducing the memory usage.

#59