feat(leader): Allow leader entries to specify context requirements.
Now Leader entries also have an optional `context` (`[]string`)
attribute. When present, it specifies the required context keys
that MUST be available for the key to be shown.

This is useful for example for keys or submenus that must be shown
only under certain contextual conditions.

The following example restricts the `e` key to be available ONLY when
both `$file` and `$change_id` are available in context. That is, when
selected a change file.

```toml
[leader.e]
context = [ "$file", "$change_id" ]
help = "Edit current file"
send = [ "$", "jj edit $change_id && $EDITOR $file", "enter" ]
```