feat(revisions): K jumps to revision first child.
This change adds `K`, analogous to `J` for jumping
to revision child (up in the revisions graph).

This is helpful when moving on non-linear revisions graphs.
Seems to work correctly when a revision more than one child
moving correctly to the intuitive one.

On help page, `J`, `K` and `@` movement keys are shown in a single row
to save some vertical space.