tests: delay quitting the shell to give enough time to clear messages in the queue
this is to prevent the race condition when the last action of the model closes and also issues jj command run. In some cases, close was processed before the jj command run and caused test to fail as the expected command was not run (didn't have a chance to run).