fix(ace_jump): Correctly handle highligting for ambigous change_ids.
Having two revisions with change_ids like: `XYZ` and `X`.


When ace_jump is activated, `X` is highligted as aceKey on both revisions.
Pressing `x` correctly moves the aceKey on first revision to be `Y` but
while the second one remains correct with an `X` key-binding, the
renderer was no longer highligting `X`. This PR fixes highligting of
short ids when they are still ambigous.


A future improvement could be for `XYZ` and `X` revisions, make `Y` to be the
aceKey for the first revision and `X` the aceKey for the second. However this
must also take into account other possible revisions like `YXZ`, etc.

However, our current approach seems comfortable since the first character is
_always_ selected as the first aceKey. And only in cases where ambiguity
is present like having two revisions prefixed with `X`, (like `XYZ` and `X`),
then disambiguation happens on the second stroke (either `Y` or `X` again)